requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9
Django==3.0.5
ecdsa==0.15
pyasn1==0.4.8
pycparser==2.20
python-dotenv==0.13.0
python-jose==3.1.0
pytz==2019.3
rsa==4.0
six==1.14.0
sqlparse==0.3.1
